changed the datatype int to double in addition class

+++ Calculator Assignment/src/com/prajwal/Addition.java
@@ -3,13 +3,13 @@ package com.prajwal;
 import java.util.ArrayList;
 
 public class Addition {
-    ArrayList<Integer>numbers=new ArrayList<>();
+    ArrayList<Double>numbers=new ArrayList<>();
 
-    public Addition(ArrayList<Integer>values){
+    public Addition(ArrayList<Double>values){
         numbers=values;
     }
-    public int add(){
-        int temp=0;
+    public double add(){
+        double temp=0;
         for(int i=0;i<numbers.size();i++){
             temp+= numbers.get(i);
         }
